Add minimal support for diff to monitor plugin
This change adds very minimal diff support for just the monitor plugin.
It only does a diff based on the arguments passed in by the user and
only supports only the "present" state.

To support a "full" diff, it would be necessary to make an additional
network call to get the new monitor configuration from the instance. To
avoid this, the code just diffs the changes passed in.